Edits on the comment moudle & commentsCounter
