--- src/modules/testFuncs.js
@@ -1,3 +1,0 @@
-const commentsCounter = (comments) => comments.length;
-
-export default commentsCounter;
